elf: Move Elf64ShdrIterator to section_headers.rs
It logically belongs to the section header implementation, so move it
there.
